fix merge issues with pubsub
